Create lister users
This creates the table and model for the first type of user the app will have.

A lister will be estate agents or other clients who will login to the site to
upload properties.

Our first priority is to get content on the site so that when we go public new
visitors have something to use.

Future models will be admins and users looking for property.